Download a series as cbt
Sometimes the ad/group notice isnt absolute and will be missed.
Chapter and title names come from the site, pages are numbered 000.jpg
Downloads to a folder with the title name.

May break at anytime. May be a violation of certain sites terms of
service (IANAL), don't if you don't have the rights to.